worker/requirements.txt
fastapi==0.115.8
langchain==0.3.19
langchain-community==0.3.18
numpy==2.2.3
openai==1.64.0
pandas==2.2.3
pinecone==6.0.1
pymupdf==1.25.3
python-dotenv==1.0.1
python-multipart==0.0.20
sqlalchemy==2.0.38
tiktoken==0.9.0
uvicorn==0.34.0
xmltodict==0.14.2
